feat(release): fallback to predefined CI/CD variable for default branch
This change set adds a fallback to the predefined CI/CD variable when
trying to determine the default branch of a project when a release
should be created without an explicitly specified `ref` and without
access to the Projects API.

That is the case when running in CI with a CI Job Token that doesn't
have access to that particular API, but releasing from CI is a very
common use case.

This is especially useful with the newly introduced
`GLAB_ENABLE_CI_AUTOLOGIN` feature.